moving contests api hook
