+++ arlo-client/src/components/Audit/Setup/Contests/index.tsx
@@ -21,7 +21,7 @@ import FormButton from '../../../Atoms/Form/FormButton'
 import { IContests } from './types'
 import schema from './schema'
 import { ISidebarMenuItem } from '../../../Atoms/Sidebar'
-import useContestsApi from './useContestsApi'
+import useContestsApi from '../useContestsApi'
 import { IContest, ICandidate } from '../../../../types'
 import { api } from '../../../utilities'
 import DropdownCheckboxList from './DropdownCheckboxList'
